% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{construct_league_url}
\alias{construct_league_url}
\title{Construct URL for a league's page}
\usage{
construct_league_url(league_name, year)
}
\arguments{
\item{league_name}{League name as seen on understat}

\item{year}{Year (or season) as seen on understat}
}
\description{
Help with constructing the URL of a league's page
on understat, for a given season (year).
}
\examples{
\dontrun{
construct_league_url(league_name = "EPL", year = 2018)
}
}
\keyword{internal}
