% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{construct_team_url}
\alias{construct_team_url}
\title{Construct URL for a team's page}
\usage{
construct_team_url(team_name, year)
}
\arguments{
\item{team_name}{Team name as seen on understat}

\item{year}{Year (or season) as seen on understat}
}
\description{
Help with constructing the URL of a team's page
on understat, for a given season (year).
}
\examples{
\dontrun{
construct_team_url(team_name = "Newcastle United", year = 2018)
}
}
