% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{get_league_meta}
\alias{get_league_meta}
\title{Get information on currently supported leagues}
\usage{
get_league_meta()
}
\description{
Retrieve useful metadata on leagues currently supported by understat.
}
\examples{
\dontrun{
get_league_meta()
}
}
