% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{get_league_seasons}
\alias{get_league_seasons}
\title{Get information about a leagues seasons}
\usage{
get_league_seasons(league_name)
}
\description{
Retrieve seasons on understat for a given league.
}
\keyword{internal}
