% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{get_league_teams_data}
\alias{get_league_teams_data}
\title{Get team stats for a league's roster}
\usage{
get_league_teams_data(league_url)
}
\arguments{
\item{league_url}{League's understat url}
}
\description{
Retrieve team data for a league listed on understat.
}
\examples{
\dontrun{
get_league_teams_data(league_url = "https://understat.com/league/EPL/2018")
}
}
