% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{get_league_teams_stats}
\alias{get_league_teams_stats}
\title{Get stats on a league's roster of teams}
\usage{
get_league_teams_stats(league_name, year)
}
\arguments{
\item{league_name}{League name}

\item{year}{Year (season)}
}
\description{
Retrieve data for all teams in a league season listed
on understat.
}
\examples{
\dontrun{
get_league_teams_stats(league_name = "EPL", year = 2018)
}
}
