% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{get_league_urls}
\alias{get_league_urls}
\title{Get URLs for supported leagues}
\usage{
get_league_urls()
}
\description{
Retrieve URLs for leagues currently supported by understat.
}
\examples{
\dontrun{
scrape_league_urls()
}
}
