% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{get_leagues_meta}
\alias{get_leagues_meta}
\title{Get information on available leagues and seasons}
\usage{
get_leagues_meta()
}
\description{
Retrieve useful metadata on leagues currently supported by understat.
}
\examples{
\dontrun{
get_leagues_meta()
}
}
