% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{get_match_stats}
\alias{get_match_stats}
\title{Get player stats for a given match}
\usage{
get_match_stats(match_id)
}
\arguments{
\item{match_id}{Match ID}
}
\value{
a tibble
}
\description{
Retrieve data for all players in a match listed on understat.
}
\examples{
\dontrun{
get_match_stats(match_id = 11662)
}
}
