% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{get_player_season_stats}
\alias{get_player_season_stats}
\title{Get season-level stats for a player}
\usage{
get_player_season_stats(player_id)
}
\arguments{
\item{player_id}{Player ID}
}
\description{
Retrieve season-level data for a player
listed on understat.
}
\examples{
\dontrun{
get_player_season_stats(player_id = 882)
}
}
