% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{get_player_shots}
\alias{get_player_shots}
\title{Get shotstats for a player}
\usage{
get_player_shots(player_id)
}
\arguments{
\item{player_id}{Player ID}
}
\value{
a tibble
}
\description{
Retrieve shot data for a player listed on understat.
}
\examples{
\dontrun{
get_player_shots(player_id = 882)
}
}
