% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{get_team_players_data}
\alias{get_team_players_data}
\title{Get player stats for a team's roster}
\usage{
get_team_players_data(team_url)
}
\arguments{
\item{team_url}{Team's understat url}
}
\description{
Retrieve player data for a team listed on understat.
}
\examples{
\dontrun{
get_team_players_data(team_url = "https://understat.com/team/Newcastle_United/2018")
}
}
