% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{scrape_league_urls}
\alias{scrape_league_urls}
\title{Get URLs for supported leagues}
\usage{
scrape_league_urls()
}
\description{
Retrieve URLs for leagues currently supported by understat.
}
\examples{
\dontrun{
scrape_league_urls()
}
}
