% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understatr-package.R
\docType{package}
\name{understatr}
\alias{understatr}
\alias{understatr-package}
\title{Get Tidy understat Data}
\description{
Provides tools to retrieve understat \url{https://understat.com/}
data in a tidy format.
}
\details{
\itemize{
\item URL: \url{https://github.com/ewenme/understatr}
\item BugReports: \url{https://github.com/ewenme/understatr/issues}
}
}
\author{
Ewen Henderson (ewenhenderson@gmail.com)
}
