% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{get_league_seasons}
\alias{get_league_seasons}
\title{Get information on a league's seasons}
\usage{
get_league_seasons(league_name)
}
\description{
Retrieve available seasons for a given league on understat.
}
\keyword{internal}
